move function to utils; patch test
