+++ src/agoradatatools/process.py
@@ -4,7 +4,6 @@ from typing import Optional, Union, Dict, Any, List
 import inspect
 import synapseclient
 from pandas import DataFrame
-import pandas as pd
 from typer import Argument, Option, Typer
 
 from agoradatatools.errors import ADTDataProcessingError, ADTDataValidationError
@@ -165,7 +164,7 @@ def process_dataset(
             dataset_name=dataset_name,
             dataset_obj=dataset_obj[dataset_name],
         )
-        df = convert_transformation_result_to_dataframe(result, dataset_name)
+        df = utils.convert_transformation_result_to_dataframe(result, dataset_name)
 
     else:
         df = entities_as_df[list(entities_as_df)[0]]
@@ -381,33 +380,6 @@ def process_all_files(
         )
 
     reporter.update_table()
-
-
-def convert_transformation_result_to_dataframe(
-    result: Union[DataFrame, dict, list, Any], dataset_name: str
-) -> DataFrame:
-    """Convert the result of a custom transformation to a pandas DataFrame.
-
-    Args:
-        result: The result from a custom transformation function
-        dataset_name (str): Name of the dataset for error reporting
-
-    Returns:
-        DataFrame: The converted result as a pandas DataFrame
-
-    Raises:
-        TypeError: If the result type is not supported for conversion
-    """
-    if isinstance(result, DataFrame):
-        return result
-    elif isinstance(result, dict):
-        return DataFrame.from_dict(result)
-    elif isinstance(result, list):
-        return pd.DataFrame(result)
-    else:
-        raise TypeError(
-            f"Custom transformation for dataset {dataset_name} returned an unsupported type: {type(result)}."
-        )
 
 
 app = Typer()
